requirements.txt
gradio==4.15.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless==4.8.1.78
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0
realesrgan>=0.3.0
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.8
